refactor: rename _injection_enums.py to injection_enums.py and update imports
