+++ .gitignore
@@ -25,3 +25,4 @@ venv
 /docker_compose_files/core/logs/
 /docker_compose_files/core/configuration/generated/
 /.editorconfig
+/docs/superpowers
