+++ src/quickapp/common/synthetic_injection/injection_enums.py
