+++ src/quickapp/timestamp_tooling/_timestamp_injection_transformer.py
@@ -1,7 +1,7 @@
 from aidial_sdk.chat_completion import Message
 from injector import ProviderOf, inject
 
-from quickapp.common.synthetic_injection._injection_enums import (
+from quickapp.common.synthetic_injection.injection_enums import (
     InjectionFrequency,
     InjectionPosition,
 )
