+++ src/tests/unit_tests/common/test_staged_tool_synthetic_injector.py
@@ -5,7 +5,7 @@ import pytest
 from aidial_sdk.chat_completion import Message, Role
 
 from quickapp.common.completion_result import CompletionResult
-from quickapp.common.synthetic_injection._injection_enums import (
+from quickapp.common.synthetic_injection.injection_enums import (
     InjectionFrequency,
     InjectionPosition,
 )
