+++ src/tests/unit_tests/common/test_synthetic_tool_call_injector.py
@@ -1,7 +1,7 @@
 import pytest
 from aidial_sdk.chat_completion import Message, Role
 
-from quickapp.common.synthetic_injection._injection_enums import (
+from quickapp.common.synthetic_injection.injection_enums import (
     InjectionFrequency,
     InjectionPosition,
 )
